moved reading files to common class
